refactor(generate-g6): make generate.rs orchestrator-only
